old/REST API/requirements.txt
blinker==1.6.2
click==8.1.3
Flask==2.3.2
grpcio==1.54.0
grpcio-tools==1.54.0
importlib-metadata==6.6.0
itsdangerous==2.1.2
Jinja2==3.1.2
joblib==1.2.0
MarkupSafe==2.1.2
numpy==1.24.3
pandas==2.0.1
protobuf==4.22.4
python-dateutil==2.8.2
pytz==2023.3
scikit-learn==1.0
scipy==1.10.1
six==1.16.0
threadpoolctl==3.1.0
tzdata==2023.3
Werkzeug==2.3.3
xgboost==1.4.0
zipp==3.15.0
